管理者セッション確認機能追加

+++ TECH.I.S_laravel/app/Http/Controllers/EventController.php
@@ -15,13 +15,21 @@ class EventController extends Controller
      */
     public function index(Request $request)
     {
-        $Event = Event::where('date','2020-10-30')->where('administrator_id',1);
+        // セッション確認
+        $administrator_id = session()->get('admin_id');
+        if(empty($administrator_id)){
+            return redirect('/admin');
+        }
+
+        // イベント情報取得
+        //$Event = Event::where('date','2020-10-30')->where('administrator_id',$administrator_id);
+        $Event = Event::where('date','2020-10-30');
         if ($Event === null) {
             $EventResponse = array();
         }else{
-        $Event = Event::where('date','2020-10-30')->where('administrator_id',1)->get()->toArray();
+        //$Event = Event::where('date','2020-10-30')->where('administrator_id',$administrator_id)->get()->toArray();
+        $Event = Event::where('date','2020-10-30')->get()->toArray();
         $EventResponse = $Event;
-        //print_r($EventResponse); exit;
         }
         return view('event_input', [
             'Events' => $EventResponse,
@@ -36,6 +44,13 @@ class EventController extends Controller
      */
     public function edit(Request $request)
     {
+        // セッション確認
+        $administrator_id = session()->get('admin_id');
+        if(empty($administrator_id)){
+            return redirect('/admin');
+        }
+
+        // バリデーション
         $this->validate($request, [
             'event_name' => 'required',
             'start_time' => 'required',
@@ -45,14 +60,14 @@ class EventController extends Controller
             return redirect('event_input')->with('flash_message', '終了時刻は開始時刻より後に設定してください');
         }
         else{
+        // イベント登録
         Event::create([
-            'administrator_id' => 1,
+            'administrator_id' => $administrator_id,
             'body' => $request->event_name,
             'date' => "2020-10-30",
             'start_time' => date('g:i',strtotime($request->start_time)),
             'end_time' => date('g:i',strtotime($request->end_time))
         ]);
-        //メモがあれば更新
         return redirect('event_input')->with('flash_message', '登録が完了しました');
         }
     }
@@ -65,6 +80,11 @@ class EventController extends Controller
      */
     public function delete(Request $request, $id)
     {
+        // セッション確認
+        $administrator_id = session()->get('admin_id');
+        if(empty($administrator_id)){
+            return redirect('/admin');
+        }
         Event::find($id)->delete();
         return redirect('event_input')->with('flash_message', '削除が完了しました');
     }
